requirements.txt
eel
requests
speedtest-cli